refactor: 패키지 분리 및 컴포저블 분리
