+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/compose/edit/personal/rename/component/BottariRenameButton.kt
@@ -1,4 +1,4 @@
-package com.bottari.presentation.compose.edit.personal.component
+package com.bottari.presentation.compose.edit.personal.rename.component
 
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.wrapContentHeight
